requirements.txt
gradio>=5.0.0
python-dotenv
langchain
langchain-ollama
langchain-community
langchain-google-genai
langchain-text-splitters
tavily-python
faiss-cpu
pydantic
